requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.34.0
Babel==2.12.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
darkdetect==0.8.0
discord.py @ git+https://github.com/rapptz/discord.py@064cbd11254e0862e4637093b969cf6cdb020608
Eel==0.16.0
frozenlist==1.3.3
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.2
idna==3.4
keyboard==0.13.5
multidict==6.0.3
numpy==1.25.1
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
pefile==2023.2.7
Pillow==9.5.0
pycparser==2.21
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pynput==1.7.6
pyparsing==3.0.9
PySimpleGUI==4.60.4
pytube==15.0.0
pywin32-ctypes==0.2.0
requests==2.28.2
schedule==1.2.0
six==1.16.0
thumbnail==1.5
tkcalendar==1.6.1
tkinter-tooltip==2.1.0
urllib3==1.26.15
whichcraft==0.6.1
yarl==1.8.2
zope.event==4.6
zope.interface==6.0
